+++ apps/website/src/app/browse/layout.tsx
@@ -1,4 +1,4 @@
-import { Footer } from "@/app/_components/footer"
+import { Footer } from "@/components/footer"
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
